+++ CHANGELOG.md
@@ -10,9 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added automatic hangup logic to the Twilio serializer. This feature hangs up
-  the Twilio call when an `EndFrame` is received. It is enabled by default and
-  is configurable via the `auto_hang_up` `InputParam`. To use this feature, set
-  up the following env vars: `TWILIO_ACCOUNT_SID` and `TWILIO_AUTH_TOKEN`.
+  the Twilio call when an `EndFrame` or `CancelFrame` is received. It is
+  enabled by default and is configurable via the `auto_hang_up` `InputParam`.
 
 - Added `SmartTurnMetricsData`, which contains end-of-turn prediction metrics,
   to the `MetricsFrame`. Using `MetricsFrame`, you can now retrieve prediction
@@ -84,6 +83,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed an issue where LLM input parameters were not working and applied correctly in `GoogleVertexLLMService`, causing
   unexpected behavior during inference.
+
+### Other
+
+- Updated the `twilio-chatbot` example to use the auto-hangup feature.
 
 ## [0.0.63] - 2025-04-11